Feel free to throw out this code if it's unsatisfactory! Now is the best time to do so!
Here's a sort of skeleton Core, as a PartModule. It dispatches the PartModule callbacks to a list of ComputerModules.
It doesn't do anything else. I've taken all the autopilot functionality out of the Core and put it into
AttitudeController, ThrustController, StagingController, and WarpController, which are ComputerModules in their own right.

I added a DisplayModule class but haven't created any DisplayModules. See if the interface is satisfactory.

There is no GUI yet.

This code is completely untested except for the attitude controller, which seems to work (see the simple 
test of it I have implemented in MechJebCore.OnUpdate)

I turned some methods in MuUtils into Vector3 extension methods and moved them to MathExtensions; feel free
to reverse this. 

I added PartExtensions with a few methods; there will be more.

git-svn-id: http://svn.mumech.com/MuMechLib/trunk@9 fef01e1f-1acb-4511-8d79-f4a48d001ba2