Added support for sharing of info windows.
Info windows now have a "Copy" button that copies their configuration to the system clipboard. On ctrl-v MJ looks on the system clipboard for a window configuration and if it finds one uses it to make a new custom window. The configuration strings are human readable and could be e.g. swapped on the forums.

The code to access the system window from within Unity is a bit of messy black magic I found on the Web. Unfortuantely it doesn't seem like we can use the System.Windows.Forms.Clipboard class so we have to do this workaround.