Low pass filter for roll correction output
Added an additional low pass filter to reduce the noise coming out of
the roll correction PID controller.  Ideally, the input should be
smoothed, but the noise was just too much for a single low pass filter
to handle and much of the noise was originating from the roll control
output.