Fixed a bug causing unexpected behavior with Kill HS in the Translatron module.
The bug made it so that once turned checked on, attitude control could not be turned off without switching to
TMode.OFF *with* Kill HS still checked off. This has been fixed, but the fix is a bit hack-ey because I was
conserned that fixing it more generically would disrupt attitude control of other TModes, if some besides OFF and
KEEP_VERTICAL have attitude requirements. If someone knows this was unnessesary, or can think of a better solution,
feel free to build on this.