Makefile: get binary paths from user's $PATH
After weighing the pros and cons of the two, it seems that it would be
somewhat better to have these be found via the $PATH variable, rather
than full paths being specified. I don't know if people will like this,
or if it's even a good idea (some may still have to specify the
variables anyway), but It's going into a spearate commit.