Update MechJebModuleDockingGuidance.cs
Ever try to build something truly monumental in orbit? Lots of docking means a docking AP is a must. Unfortunately, the MJ AP doesn't play nicely with the built-in KSP docking port targeting system. Specifically, if the MechJebModuleDockingAutopilot.targetSize is near or larger than the ~200m maximum distance for targeting a docking port, the AP always loses its lock as it backs up to its starting distance. This update (and the update to MechJebModuleDockingAutopilot.cs) allow an override for the auto-calculated targetSize to get around this problem. Get close to the docking port, override the targetSize, and watch the docking AP do its thing!