low altitude landing safety tweaks
if we press one of the landing buttons when we're below the deceleration
burn altitude bad things can happen.

* if we're off course by > 150m then we get stuck in CourseCorrection and
  will reliably lithobrake
* i've seen craft get stuck in "coasting towards deceleration burn" and
  lithobrake but replication is unreliable

to address both cases, add shortcuts so that we wind up in the logic
in DecelerationBurn that selects KillHorizontalVelocity or FinalDescent
correctly.

also I've had two craft lithobrake at 50x warp, which is not reliably
reproducable and happens too fast to see what mode its in.  it may be
due to the 10-seconds-to-impact extrapolation not taking into account
gravitational acceleration on very low initial velocities and low
heights.