Use ship altitude instead of final alt for ascent (#779)
Now that we're continuously updating the heading based on the vessel
position, it makes more sense to use the vessel's altitude as the guess
for our direction rather than the final orbital velocity.

This trades some wildly inaccurate numbers for some slightly inaccurate
numbers.

If you want to ascend directly to a large apoapsis with a large
inclination then MJ will try to shoot for the horizontal velocity of
the apoapsis while its still on the ground.  At the absurd limit if
you punch in the SOI limit of the body you're lifting off of you'll
be asking to putter along at a few m/s and that is definitely
insufficient for a launch velocity.

This algorithm gets it a little bit wrong, but it will self correct as the
rocket ascends and it continually updates.

As a refinement we could go back to injecting the target altitude again
and then calculating the velocity at the periapsis of an orbit with the
periapsis at the surface and the apoapsis at the desired height, but i'm
not sure we need to and i haven't quite convinced myself that would be
any more correct.