launch inclination fixes for Kerbin
this patch is a bit of hackery to make two-stage burns work correctly
on Kerbin, also tested on Mun and even Gilly.

on Earth we don't need this so much because we tend to do single burns
and the Earth is really big.

the fracDV calculation essentially selects .85 as the frac to apply
to Kerbin while fading out to a correct of .00 on Earth.

the correction factor also fades out as we launch and the deltaHV
value trends down.  that means we correct stronger sooner which
should(?) reduce the steering losses from high inclination launches.

as the deltaHV value whips around we stop doing this calculation as
well (which is fairly important because deltaHV can get large again but
point in the wrong direction under some conditions).

the 0.85 value is frighteningly large, but i had to do that to get
rockets to launch correctly on Kerbin.  i can't really explain it with
numberical arguments but the rockets launch well.  on Kerbin with 175m/s
of coriolis velocity at KSC it only costed that plus an extra 16m/s of
steering losses to hit a 90.01 inclination orbit compared to a 0 degree
one.

tested with 90 degrees and 135 degrees inclination on Kerbin with SLTs
ranging from 1.1 to over 2.5.  on Gilly with a 2000 launch TWR it didn't
work because that's just insane -- one puff and you launch whichever
direction you were pointed on the surface -- but after clamping the
throttle to 1% it worked fine.

since we track continuously as we launch now any inaccuracy in the
initial guess will get corrected over time.

rockets with truly insanely high TWRs and super-short takeoff burns to
raise the apoapsis will still not work well.  the two possible
strategies there will be to simply throttle down -- or else i think we
could fade this algorithm out as the TWR gets stupidly big.  not sure
how much it matters though since at extreme TWR attitude control doesn't
help you any and you launch like a gun based on whatever your attitude
was due to the surface contour.