Use Gooding's instead Battin's Lambert Solver in Adv Transfer (#834)
2x to 2.5x speed improvement.  Should be a sizable accuracy
improvement -- Gooding claims 13 significant figures of accuracy
for the 3 rounds in this algorithm.

The fixed 8 rounds of ComputeKsi() in the Battin Solver and
seeding the NewtonSolver with a fixed guess of 10 probably gave
poor performance/accuracy/instability for certain inputs in the
old algorithm.  Perhaps the old implementation could be improved,
but this one should beat it overall (I tried fixing the old
algorithm according to Battin's recomendations in his book but
i broke it).

This algorithm could probably be cut down to 2 rounds and 5-6
significant figures to speed up porkchop plots even further.  IDK
if any KSP-uses really need the accuracy of 3 rounds at all.

I deliberately didn't do the work to fully port all the Lambert
use cases over to this yet so it can be playtested a little bit,
and to keep it easy to roll back the edited lines in the
TransferCalculator to revert the behavior back in case I screwed
something up here.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>