improve transfer optimization (#836)
updates the optimizer that picks transfer burns to be more
accurate.

the biggest change here is in not fixing the ending UT of
the transfer and looking for the closest match at that
UT.  i could have taken the two orbits and instead called
KSPs functions to search for the closest distance.  since
we're using an N-dimensional optimizer, though, i just added
the ending UT as another variable.

since constraining the UT was likely constraining the deltaV
of the solutions it seemed prudent to add the deltaV in as a
quantity to minimize.  the factor of 10,000 was chosen through
a bit of playtesting -- 100,000 was clearly too high.

with these tweaks accuracy of the transfer orbits have improved
greatly and the algorithm can generally nail the intercept, often
with 10 dV or so saved over the best value on the porkchop plot.

there's still issues with interactions.  kerbin-to-moho transfers
often get mun enounters and are somewhat poor.  kerbin-to-jool is
often perturbed but the solution generally seems to wind up in the
inner joolean moons.  kerbin-to-dres, kerbin-to-eeloo and
kerbin-to-moho are usually spot on as long as the mun doesn't get
in the way.

it seems like it should be possible to use the existing work around
that bumps the solution to the next orbit and restarts the solver in
order to avoid encounters, but i don't quite have a perfectly clear
picture in my head of how to do that in a generic way.

more useful would probably be to include a button on the porkchop
planner to include the destination capture DV in the plot and
selection.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>