Improving the HeadingForLaunchInclination function. (#851)
The original version of this code had to transition between the ground frame of reference and the inflight frame of reference, so it had a complex reserve system to figure out the transition.
Then the code was modified to use the vessel's frame of reference, so the transition is no longer needed.  The current code now calculates the difference between your target velocity vector, and your current velocity vector,
and returns the heading for that.  That's the exact heading we want to take (if we were to do an impulse)