Add actual bugfix for ullaging engines that aren't firing
I'm not sure why I originally wrote it this way, but I suspect
that it made sense before some kind of rearrangement of the
ThrustController.  Now what is happening is that the targetThrottle
logic has been moved to the top of the controller, and there's
a condition where the targetThrottle could be set but the
Users are == 1.  This would happen, I think, if Ascent Guidance
was firing the throttle and had targetThrottle = 1.0 but then
the user disengaged it, it removed itself as a user, but left
the targetThrottle set.  In that case nothing is actually
firing the throttle (s.mainThrottle == 0) but the ullage code
is firing anyway.  There's no later reference to targetThrottle
after the ullaging code so its pointless to be looking at
anyway.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>