2 Ullage limiter fixes (#883)
* fix ullage limiter condition

this mostly just fixes it so that when you've got unstable ullage
its not constantly green until you actually try to throttle and
it gets prevented.

also if you had been limited to zero by some other condition and
had unstable ullage, the unstable ullage condition would have
"overwritten" the other throttle limiter.

believe this is only a cosmetic bug in the UI.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>

* Add actual bugfix for ullaging engines that aren't firing

I'm not sure why I originally wrote it this way, but I suspect
that it made sense before some kind of rearrangement of the
ThrustController.  Now what is happening is that the targetThrottle
logic has been moved to the top of the controller, and there's
a condition where the targetThrottle could be set but the
Users are == 1.  This would happen, I think, if Ascent Guidance
was firing the throttle and had targetThrottle = 1.0 but then
the user disengaged it, it removed itself as a user, but left
the targetThrottle set.  In that case nothing is actually
firing the throttle (s.mainThrottle == 0) but the ullage code
is firing anyway.  There's no later reference to targetThrottle
after the ullaging code so its pointless to be looking at
anyway.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>