Make thrust controller default to 1.0 under NaN conditions
I don't see anywhere in the code where these values would ever be NaN,
but it makes more sense to disable the limiter and allow throttle under
these conditions.

If we don't then the computed accelleration of the ship will be zero and
the burntime will be infinite which greatly throws off things like
the node executor.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>