fix low inclination ascent wiggles
when launching from Kennedy (28.608) to the moon (28.363) the
ascent guidance would start (correctly) by burning due East (can't
launch to lower than your launch latitude).

as the craft followed a great circle track, the latitude of the craft
would fall allowing smaller latitudes.  however this resulted in the
craft attempting an inclination change at the end of the burn by
turning north (the opposite of a southwards jog at the start of the
burn) which was both expensive and totally screwed up plane targeting.

this fixes this by just not allowing inclinations which are below the
launch latitude (also allowing inclinations which are below the current
orbital inclination which would be fine if someone winds up in that
situation through some kind of button mashing).

this makes south-jogs very hard to do, but it results in very precise 0.25
degree errors when launching to the moon from KSC in RSS.  scanning
through NK's videos from 1.1.3-era he seems to have done no better than
this and often as bad as 0.40+ when launch to the moon with a south
jog.  i think mostly his south-jogs were fighting with the old MJ ascent
guidance which did no great-circle tracking.

adding south jogs would be hard since you need to 'know' the south jog
happened and allow the permissible launch inclination to lower.  it
might be necessary at that point to allow unclamped launch inclinations
and let people jog south and then correct, but i had a hard time getting
any better results than just with the clamped inclination.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>