fix 1kN thruster suicide burn calc
i'm not sure if real fuels changed its API so that this no longer works
correctly or if it was not correct in the first place -- but this
definitely gets minthrust/maxthrust very wrong for the 1kN thruster
in realism overhaul / real fuels.

also applied the thrust limiter w/o the ullage clipping to zero to the
max thrust, which is what i thought this problem was going to be in the
first place -- this still looks like more correct code here, but wasn't
the root of the bug i was seeing in the suicide calcs.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>