Add freeMolecularAerothermalFlux to VesselState
Used to determine when fairings can be jettisoned.

Only free molecular aerothermal flux is calculated instead of
aerothermal flux in general because only the former is relevant for
determining when to drop fairings [0].

Display just "Aerothermal flux" because as much as I like "Free
molecular aerothermal flux" for its accuracy, it's just too long.

    [0]: ferram4 in #RO: "...the fairings aren't going to be dropped
         below the point where free molecular flow is happening"