More transfer calculator fixes
the major point of this fix is actually the removal of the line

orbit.StartUT = t

which seems to cause issues.  that was being done so that
PatchedConics.CalculatePatch() could be called with orbit.StartUT
consistently, but i unrolled the loop so it gets called with 't'
first and then orbit.StartUT.  more code, but avoids the hack.

then there's some successive APIs that are called:

- getTruePositionAtUT (if we're not even in the correct solar transfer
orbit somehow)
- getRelativePositionAtUT (for the solar transfer orbit)
- getRelativePositionFromTrueAnomaly(0) (aka periapsis distance once we
  get into the target SOI)
- difference in dV once we're hitting the planet pretty much exactly

since the magnitude takes sucessive jumps up at every step the optimizer
should be able to handle is and continuously trickle down to the target.

i pinned the UT_arrival time again.  i don't think it makes a lot of
sense to try to optimize that since someone may have tried to pick a
target off of the porkchop plot and they should get more or less what
they pick.

did a bunch of whitespace cleanup as well.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>