fix PID wiggles
the old code would reset the PID controller every 10 degrees a
rocket pitched over, because it remembered the last value it
reset the pid at, not the actual last value the last tick.

most of the variables here are a bit useless, the variable it sets
at the end is the last value when it starts in the setter.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>