fix "launch to plane" again again
this time with feeling.

since we want to fully match the angular momentum vector for inc, lan
and one component that sort of corresponds to half of sma/ecc or
ApR/PeR pairs fully specify that.  then fix the flight path angle.
then take only one of velocity or radius to match (the other one
is fully constrained by the magnitude of the h-vec so is not free).

then use the argp transfersality conditions.

the result is that now the optimizer stops flailing and settles down
to 8-10 iterations when converged.

also added the znorm to the status output.

i don't think this has a singularity, but we'll have to see.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>