add LogicalStageTracking
this is necessary to act as an intermediary between PVG and
the delta V stats in MJ.  it offers objects which are themselves
static that correspond to logical rocket stages even though
the ksp stage and the position in the stage stats may bounce
around as the user drags the staging crap all over the place.
of course the stats referenced by the object may change, but
the first stage of the rocket is always a given object and it
gets marked as 'staged' when it gets dropped and then the
stage manager tracks how many of those significant stages have
been dropped so we always know 'which' numbered stage we are
on and a sort of trusted 'what rocket stage number is this'.

the fact that i've spent like 1.5 years working on this and
think i may have finally more or less nailed the design of what
we really need indicates how bad of a rocket simulator KSP is.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>