use effective thrust of the stage
the start thrust can include ullage motors which burnout and
throw off the whole ∆v calculation.  this uses an "effective
thrust" over the whole stage interval based on the mass loss and
the exhaust velocity, which itself is based on the sum of the
∆v (which is accurate and summed in 'chunks') and the total
mass loss over the interval.

so what we get that is entirely accurate is the time of the whole
burn, along with the summed up ∆v, and the mass loss.  from that
effective isps and constant thrusts are computed.

really maybe what should be done is to expose those smaller chunks
in the API so that each array item in vacStats no longer corresponded
1:1 to a KSP stage.  but this is a way, way, way easier fix.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>