Display ScreenMessage about unstable ignitions
If "Prevent unstable ignitions" is checked this will now throw up a
message on the screen indicating what MJ is doing.

This may help players who are using sounding rockets, where this
setting may not be appropriate (since there is no way to RCS
ullage up a sounding rocket).

Generally the safe way to fly with MJ is to set "Use RCS to ullage"
and to unset "Prevent unstable ignitions".  The RCS setting will
only apply when there are active RCS modules that have the correct
directionality to ullage the stage, and should always DTRT while
not affecting sounding rockets and any other stages that do not
have any RCS.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>