Better PVG terminal guidance
The condition to cut the throttle for both proper engine terminal
guidance and RCS terminal guidance is now dependent upon not
overshooting the target orbital angular momentum instead of just
count down predicted time.  This is considerably more accurate and
is accurate even in the face of bugs that screw up the mass/thrust
of stage analysis.

This also surfaces those bugs through error messages on screen
during the launch, since those also need fixing.

This change also auto-enables RCS for coast periods and for
terminal guidance.  Can still be disabled if its buggy for some
reason.

Tested with launches both to plane and not and to highly elliptical
orbits and with rockets that have screwy root parts and inaccurate
∆v stats (and which wind up with very negative tgo+vgo values once
the burn completes, but the target conditions were highly accurate).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>