Fix disabling autostaging in flight
If the "sub" ascent modules (Classic, GT, PVG) are enabled they are
currently registering themselves with autostaging if autostaging is
enabled on the autopilot.  That means that in flight if autostaging
is disabled that only the code runs in the main Autopilot class
to try to disable it, but the sub-module is still registered.

This should always just have been left the responsibility of the main
autopilot class itself and not in the base class of the sub controllers.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>