Fix wobbly dV display in the last stage
We need to synchronize the stage stats into the solution, but that
synchronization only happens when kicking off an optimization.  That
means that the state that was last synchronized may have no resemblance
to the current state of the stage stats, because the user may have
arbitrarily been rearranging staging in the meantime.  So we also
need to track the time that we synchronized the stage stats.

One open question I have here is why we don't simply update the stage
stats into the solution on every tick.  I don't recall why I didn't
do that.  Might have something to do with updating information
mid-optimization and creating a race condition with the next optimizer
cycle needing a consistent view over the whole time its running?  That
seems a bit weird, but the prior solution is consumed by the running
optimizer.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>