PVG core update
fixes and improvements to lots of stuff

most visible to users is all the fixes which particularly appeared with
early russian launches not coasting.

* initial guess of coast period is longer to avoid optimizer finding
negative solutions which also exist.
* when going from infinite upper stage to finite stage + coast the
overall burntime guess is clipped to the actual burntime of the rocket
to avoid starting the optimizer way past where the rocket has burned
down to zero due to exceeding the tau of the stage.
* r0 and v0 are now also hard pinned in prep for atmospheric
backpressure effects.
* the optimization function now requests the optimizer to stop when
the znorm is low enough (making it both faster and more stable and
allowing the minimum step-size to be decreased).
* a bunch of upstream code for other targeting conditions is included
here that is not yet used (but will help keep my merge conflicts down).