Add Brent's Method and convert some binary searches
Adds Brent's method of 1-d rootfinding and converts the
usages of binary searches in the OrbitalManeuverCalculator to
use it.  Change ApR, PeR, SMA and the simple Hohmann transfer
now all use it instead of binary searching.

Alone this isn't super useful, and those functions are still
a bit hacky, but the major purpose is I want to build other
stuff that requires Brent's method.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>