Fix the ApR and SMA maneuvers
The heuristics this routine was using for min/max were a little bit
weird, so removed the binary search it did to try to bracket the
root, and instead converted the ApR and SMA searches to search for
the reciprocal value (which is a continuous function) and just let
Brent got search for the root (which has to be faster/better).

I don't think this affects PeR and I hope it doesn't affect the
Hohmann transfer stuff.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>