Add a simplified and commented `TimeToPlane` algorithm
This change replaces the previous implementation of `TimeToPlane` with a
new one that is simpler and more commented.

It also fixes a bug in the handling of `LANDifference`: because
`LANDifference` was subtracted as the very last step in the angle
computations, a positive value could cause the output of `TimeToPlane`
to be negative, leading to an immediate launch even when that was not
correct.

Also, there is now a special case for very low inclinations (<.01
degrees); since there's such a small difference between different launch
times, the function now just returns 0. The old implementation always
returned a launch point 90 degrees away when the inclination was exactly
0.