Simpler PVG stage tracking
Back to a very simple "if we lose a non-zero stage in the delta-V stats then we
dropped a stage" formula.

We no longer even bother with discriminating ullage stages and just let
PVG run on them.

The problem with the whole approach that PVG attempted to do in the past
is that the "running out of deltav in a stage" event is not going to
happen on the same tick as the actual staging event or even request.

It is all complicated by pre and post stage event delays.  And even
something as simple as a heuristic to watch for a stage that is about to
drain and then assuming the subsequent staging event / zero dV stage was
due to a staging event is complicated by Test Flight which may make a
stage suddenly go to zero when it is very far away from its planned
stage time.

This basically drops support for reconfiguration of staging during
flight pretty much entirely.  If you do that, PVG will probably wig out
and need guidance to be reset.

At this point I'm very okay with eating your rocket if you mess with
staging during ascent.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>