Only track stages right before running optimizer
Avoids updating the data in a race condition with the optimizer running
in the thread, and also means that it'll be quieted right during staging
events.

The stack tracker is also not remotely a computer module any more and
needs to be moved back into the guidance controller.

I also got real tired of the "off by 100%" lolwut.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>