More attitude control buttons
Adds the ability to completely suppress actuation on any axis.

When actuation is suppressed the BetterController doesn't windup
on that axis.

Also adds the (presently unused) ability for external modules to
set a manual target omega (to be used for spin-up).

Also re-fixes launch wobblies.  There's now small wobble that happens
when steering kicks in at 5m above the pad, but there's no windup and
we should have lots of torque by then.

I don't want to fly completely unguided until we clear the pad and
roll because asymmetric thrust vehicles will probably start to flip
on launch (they may start to flip a bit already as it is).

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>