Fix transfer planner maneuver placement
1.  Fixes yet another painful annoying inverse rotation bug in
ComputeEjectionManeuver by converting from GetStateVectorsAtUT to
getRelativePositionAtUT and getOrbitalVelocityAtUT.  I've forgotten
which way is which but the latter two agree with the coordinate system
of the rest of the code in the transfer planner.

2. Converts to using a proper NLP optimizer with a cost function and
a periapsis constraint.  There's two NLPs here, one to nudge the
intercept into the SOI if the initial guess is too far away and one
to do the periapsis constraint.  After 1.12.x is the only supported
KSP version Orbit.NextCloseApproachTime() could be used for trajectories
which do not intersect the SOI to create a smooth NLP against the
periapsis radius to combine the NLPs into a single one.

The NLP solver works under all the use cases I've tried so far, but it
is slow and beachballs for awhile.  It is also not quite as accurate as
I'd like.  There may be room for perf improvements in how the NLP
solvers in alglib are used, but its the first time I've tried them in
"production".

This may not work well at all in KSP 1.10.x on RSS/RO due to the single
precision floating point operations in the maneuver planner.  Sorta
don't care though, it all needs to be upgraded.

Bunch of unrelated cleanup.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>