Optimize VesselState
Stock fixed their Moment of Inertia and Angular Momentum calcs from VesselPrecalculate, so use those instead of recomputing
Cache the results for calling FAR's CalculateVesselAeroForces.  The environment does not change much from frame to frame, but the call is expensive.  Set thresholds on the change in parameters (velocity, altitude) and use the previous result if within bounds
TODO: thresholds declared as constant are fine for demonstration, but they should probably be moved to a config area.